updated validator and sanitized data sanitizer
Updated Validator to accept error_type parameter.
Removed trim and other 'unnecessary sanitizers' from DataSanitizer. This
avoids situations where the developer would not want the data trimmed,
like when I was working with the parsedown library implementation of
markdown. So, responsibility is now on the developer to trim their
string data.